app-admin/ego depends on pyyaml and click
A yaml parser can parse JSON files, and allows
comments to be used. YAML wins in the human-interface
department. JSON is easier to learn with a tighter
spec, and better suited for data serialization.
The idea is to use YAML for configuration
files, while json should be used to interact between
APIs.

Click (dev-python click) is a library for building CLI
interfaces and pretty much does what the current ego
and epro interfaces do, but handles some corner cases a
lot better.

Piping ego module commands is allegedly easier with click,
and strips the ANSI color codes. This irritates some Funtoo
users (including me) when reading pastebin outputs.

Also, creating a git-like command-line interface with click is
a lot easier. This is quite important for the functionality of
the refactored ego and epro codebases.